+++ package.json
@@ -2,8 +2,15 @@
   "name": "netuitive-package-cli",
   "version": "0.1.2",
   "description": "Netuitive Package Interface",
+  "main": "js/index.js",
+  "types": "js/index.d.ts",
+  "scripts": {
+    "compile": "tsc -p tsconfig.json",
+    "test": "nyc mocha",
+    "prepublish": "tsc"
+  },
   "bin": {
-    "netuitive-package": "bin/netuitive-package.js"
+    "netuitive-package": "js/bin/netuitive-package.js"
   },
   "author": "TheConnMan",
   "license": "ISC",
@@ -14,5 +21,14 @@
     "json-dup-key-validator": "^1.0.2",
     "request": "^2.81.0",
     "request-promise": "^4.2.1"
+  },
+  "devDependencies": {
+    "@types/extend": "^2.0.30",
+    "@types/inquirer": "0.0.35",
+    "@types/node": "^8.0.28",
+    "@types/request": "^2.0.3",
+    "@types/request-promise": "^4.1.37",
+    "ts-node": "^3.3.0",
+    "typescript": "^2.5.2"
   }
 }
