+++ ts/util/Logger.ts
@@ -1,0 +1,14 @@
+class Logger {
+
+  private indent: string;
+
+  constructor(indent: string) {
+    this.indent = indent;
+  }
+
+  public log(message: string) {
+    console.log(this.indent + message);
+  }
+}
+
+export default Logger;
