+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "version": "2.3.3",
+  "compilerOptions": {
+    "target": "es6",
+    "module": "commonjs",
+    "noImplicitAny": false,
+    "removeComments": true,
+    "noLib": false,
+    "preserveConstEnums": true,
+    "declaration": true,
+    "suppressImplicitAnyIndexErrors": true,
+    "outDir": "./js"
+  },
+  "types": [
+      "node"
+  ],
+  "include": [
+    "./ts/**/*.ts"
+  ]
+}
