--- util/logger.js
@@ -1,9 +1,0 @@
-function Logger(indent) {
-  this.indent = indent;
-}
-
-Logger.prototype.log = function(message) {
-  console.log(this.indent + message);
-};
-
-module.exports = Logger;
